Move authentication tests from api/receiver to api/auth
